Pass norm on parse folder
